refactor(validation): move commands out of docs
Two commands triggered by doc changes that are wired to workers should
exist under validation.
